requirements.txt
pandas>=1.3.0
pytest>=7.0.0
apache-airflow>=2.7.0
apache-airflow-providers-python>=4.0.0
python-dotenv>=0.19.0
numpy>=1.21.0
scikit-learn>=1.0.0